Es ist nun immer der Ganze Beacon sichtbar (es werden keine Teile mehr in gewissen Perspektiven ausgeblendet). Nun schiesst auch der Deactivation Tower wieder (Hatte zuvor nicht, da es ein Rundungsfehler bei der Berechnung der SPS gab, wodurch diese immer auf 0 waren)

+++ src/mygame/Entitys/DeactivationTower.java
@@ -164,7 +164,7 @@ public class DeactivationTower extends Tower{
      */
     @Override
     public double getNewSPS(int newLevel) {
-        return Math.round(newLevel/20.0*100)/100.0;
+        return Math.round((double) newLevel/20.0*100.0)/100.0;
     }
         
     /**
